Move k8s tunnel to cluster package
